Fixed textures

+++ physics-engine/texture.hpp
@@ -25,7 +25,7 @@ const color_t RED = { 0xFF0000FF };
 const color_t GREEN = { 0x00FF00FF };
 const color_t BLUE = { 0x0000FFFF };
 
-texture_t xor_texture(color_t base, int width, int height);
+texture_t xor_texture(color_t base, uint16_t width, uint16_t height);
 texture_t solid_colored_texture(color_t color);
 texture_t build_texture(std::string filename);
-texture_t noise_texture(color_t base, int width, int height, double frequency, uint32_t octave, uint32_t seed);
+texture_t noise_texture(color_t base, uint16_t width, uint16_t height, double frequency, uint32_t octave, uint32_t seed);
